Add SSE server. It was originally implemented as STDIO server and launched with mcp-proxy so multiple clients could connect to a shared instance.
Extract server logic and produce both STDIO and SSE wrapped versions of the server.

Add npm scripts for launching the SSE directly or launching the STDIO with mcp-proxy as before.
