+++ esbuild.config.cjs
@@ -1,18 +1,9 @@
 const { build } = require("esbuild");
 
 build({
-  entryPoints: ["src/index.ts"],
+  entryPoints: ["src/stdio.ts", "src/sse.ts"],
   bundle: true,
   platform: "node",
   target: "node16",
-  outfile: "dist/index.cjs",
-  external: [
-    "express",
-    "fs",
-    "path",
-    "net",
-    "stream",
-    "url",
-    // List other Node.js built-in modules you need to keep externalized
-  ],
+  outdir: "dist/"
 }).catch(() => process.exit(1));
