+++ package.json
@@ -2,15 +2,16 @@
   "name": "goose-team",
   "version": "0.3.0",
   "scripts": {
-    "build": "node esbuild.config.cjs",
-    "mcp-proxy": "npx mcp-proxy --port 3001 --endpoint /sse node dist/index.cjs",
+    "build": "node esbuild.config.cjs && echo '{\"type\": \"commonjs\"}' > dist/package.json",
+    "stdio:proxy": "npx mcp-proxy --port 3001 --endpoint /sse node dist/stdio.js",
+    "sse:direct": "cd dist && node sse.js",
     "inspector": "npx @modelcontextprotocol/inspector",
     "agent:text": "goose run -i instructions/protocol-as-text.md",
     "agent:chart": "goose run -i instructions/protocol-as-chart.md",
     "agent:wait:test": "goose run -i instructions/wait-loop-test.md",
     "format": "prettier --write ./src",
     "typecheck": "tsc --noEmit -p tsconfig.json --composite false",
-    "lint": "eslint .",
+    "lint": "eslint --ignore-pattern 'dist/*'  .",
     "lint:fix": "eslint ./src --fix",
     "test": "jest"
   },
@@ -37,7 +38,7 @@
     "typescript": "^5.7.3"
   },
   "dependencies": {
-    "@modelcontextprotocol/inspector": "^0.4.1",
+    "@modelcontextprotocol/inspector": "^0.6.0",
     "@modelcontextprotocol/sdk": "^1.5.0",
     "@types/express": "^5.0.0",
     "mcp-proxy": "^2.4.0",
